requirements.txt
aiogram>=3.3.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
yookassa>=3.0.0

